backend/requirements.txt
tornado==5.0.2
SQLAlchemy==1.1.4
alembic==0.9.9
selenium==3.12.0
tzlocal==1.5.1
pytz==2018.4
humanize==0.5.1
obspy==1.1.0 
jsonpickle==0.9.5 
apscheduler==3.5.1
python-slugify==1.2.6